+++ engine/spec/helpers/citizens_advice_components/form_builder_cads_check_box_spec.rb
@@ -24,8 +24,6 @@ RSpec.describe CitizensAdviceComponents::FormBuilder do
           checked: true
         )
 
-        pending "Bug: checkbox generates IDs by name"
-
         expect(page).to have_field(
           "example_form[confirmation]",
           with: "1",
@@ -51,25 +49,12 @@ RSpec.describe CitizensAdviceComponents::FormBuilder do
 
       it "renders error message" do
         expect(page).to have_css(
-          "[id='example_form[confirmation]-error']",
-          text: "Confirmation can't be blank"
-        )
-
-        pending "Bug: checkbox generates IDs by name"
-
-        expect(page).to have_css(
           "#example_form_confirmation-error",
           text: "Confirmation can't be blank"
         )
       end
 
       it "sets aria-describedby" do
-        expect(page).to have_css(
-          "input[aria-describedby='example_form[confirmation]-error']"
-        )
-
-        pending "Bug: checkbox generates IDs by name"
-
         expect(page).to have_css(
           "input[aria-describedby='example_form_confirmation-error']"
         )