+++ engine/spec/helpers/citizens_advice_components/form_builder_cads_collection_check_boxes_spec.rb
@@ -91,8 +91,6 @@ RSpec.describe CitizensAdviceComponents::FormBuilder do
         around { |example| I18n.with_locale(:cy) { example.run } }
 
         it "translates optional label" do
-          pending "Not yet implemented"
-
           expect(page).to have_css "legend", text: "Currency (dewisol)", normalize_ws: true
         end
       end
@@ -136,8 +134,6 @@ RSpec.describe CitizensAdviceComponents::FormBuilder do
       end
 
       it "sets aria-describedby" do
-        pending "Not yet implemented"
-
         expect(page).to have_css "fieldset[aria-describedby=example_form_currency-hint]"
       end
     end
@@ -178,14 +174,10 @@ RSpec.describe CitizensAdviceComponents::FormBuilder do
       it "renders error message" do
         expect(page).to have_text "Currency is not included in the list"
 
-        pending "Doesn't yet associate an ID with the error message"
-
         expect(page).to have_css "#example_form_currency-error", text: "Currency is not included in the list"
       end
 
       it "sets aria-describedby" do
-        pending "Doesn't yet associate an ID with the error message"
-
         expect(page).to have_css "fieldset[aria-describedby='example_form_currency-error']"
       end
 
@@ -201,8 +193,6 @@ RSpec.describe CitizensAdviceComponents::FormBuilder do
         end
 
         it "sets multiple aria-describedby" do
-          pending "Doesn't yet associate an ID with the error message"
-
           expect(page).to have_css "fieldset[aria-describedby='example_form_currency-error example_form_currency-hint']"
         end
       end